cursoservice/requirements.txt
# Runtime
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
httpx>=0.27

# (Opcional futuro: base de datos)
# sqlalchemy>=2.0
# alembic>=1.13

# Testing & calidad
pytest>=8.2
pytest-asyncio>=0.23
ruff>=0.5
black>=24.8
pymongo[srv]==4.8.0
python-dotenv==1.0.1

